auth_service_compose/auth_service/tests/functional/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
aioredis==1.3.1
pydantic==1.9.0
python-dotenv==0.20.0
pytest-html==3.1.1
psycopg2-binary==2.9.3